reddit-skills/pyproject.toml
[project]
name = "reddit-skills"
version = "0.1.0"
description = "Reddit marketing automation Skills for Claude Code"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "praw>=7.8.0",
    "google-genai>=1.67.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.isort]
known-first-party = ["reddit_skills"]

[tool.pytest.ini_options]
testpaths = ["tests"]
